Feature/tagged logging (#1)
* Add tagged logging support

Implement tagged logging similar to ActiveSupport::TaggedLogging.

* Use log_tags attribute on logger

Don't use the default_options to pass the tags into the logger.  This
causes the tags field to be sent over to graylog as an array.  Instead
pass the tag names into an attribute called log_tags that matches with
the Rails standard of config.log_tags.

* Replace Thread#[] with Thread#thread_variable_get/set

* Skip log tags tests for Ruby 1.9

* Fix CI for JRuby

Co-authored-by: Mark Glenn <mark.glenn@mbx.com>
Co-authored-by: Marcus Ilgner <mail@marcusilgner.com>
Co-authored-by: Vladimir Kitaev <vkitaev@freeconferencecall.com>